Добавляет модуль с картой
