+++ js/main.js
@@ -1,3 +1,4 @@
 import './card.js';
 import './form.js';
 import './form-validate.js';
+import './map.js';
